ki_metadata_extended/requirements.txt
fastapi
uvicorn
pillow
transformers
torch
deepface
opencv-python-headless
neo4j
tf-keras
python-multipart
streamlit
plotly
boto3
